Money, SwingCurrencyDialog and SwingMoneyDialog updated

+++ src/main/java/software/ulpgc/moneycalculator/Money.java
@@ -1,6 +1,6 @@
 package software.ulpgc.moneycalculator;
 
-public record Money(long amount, Currency currency) {
+public record Money(double amount, Currency currency) {
     @Override
     public String toString() {
         return amount + " " + currency;
